+++ common/src/main/java/dev/aika/smsn/config/SMSNConfig.java
@@ -79,6 +79,8 @@ public class SMSNConfig extends ModConfig {
     @Category("qol")
     @LoaderSpecific(LoaderType.FORGE)
     public boolean immersiveCavesDiscordMessage = false;
+    @Category("qol")
+    public boolean ribbitsNoMorePrideMonth = false;
 
     @MixinList(value = @Category("mixins"), prefix = "dev.aika.smsn.")
     public Set<String> disabledMixins = new HashSet<>();